+++ examples/coagents-shared-state/agent/translate_agent/agent.py
